Cleaned Code. Moved Home to another page. Redesigned map and made it focus on buildings. Added Size
