+++ webapp/application.py
@@ -1,9 +1,9 @@
 from dash import Dash, dcc, html, Input, Output, callback
 import dash_bootstrap_components as dbc
-from pages import navbar, home, page1, page2
+from pages import AboutDataSet, navbar, home, page2
 import os
 
-cdn_Style_Sheets = [dbc.themes.BOOTSTRAP]
+cdn_Style_Sheets = ['https://bootswatch.com/5/simplex/bootstrap.min.css']
 
 cdn_Scripts = []
 
@@ -27,8 +27,8 @@ Dash_App.layout = html.Div([
 @callback(Output('page-content', 'children'),
               Input('url', 'pathname'))
 def display_page(pathname):
-    if pathname == '/page1':
-        return page1.layout
+    if pathname == '/AboutDataSet':
+        return AboutDataSet.createLayout()
     elif pathname == '/page2':
         return page2.layout
     else:
